Add files via upload
Arduino connected to potentiometer and when turn potentiometer sends a value to python via serial port.  And then in Python with VPython the code displays a Volt Meter.  At top of this code is the file to save as .ino and below that save it as .py and delete or comment out the .ino part.